requirements.txt
fastapi[all]
psycopg2-binary
pydantic-settings
sqlalchemy